Module Colis

CoLiS-Language

module Internals = Colis_internals

type colis = Colis__.Syntax__Syntax.program

The abstract syntax of CoLiS programs.

module Language : sig ... end

The CoLiS language

module Common : sig ... end

Modules shared between the different interpreters

module Concrete : sig ... end

The concrete interpreter

module FilesystemSpec : sig ... end

Specification of the initial filesystem

module Constraints = Colis_constraints

The interpreters

type sym_config = {

loop_limit : int;
Maximum number of iterations of while loops in symbolic execution

stack_size : int;
Maximum height of the call stack in symbolic execution

filesystem_spec : FilesystemSpec.t;
Specification of the initial filesystem

}

module SymbolicConstraints : sig ... end

The symbolic interpreter using constraints on the mixed backend of SymbolicUtility

module SymbolicTransducers : sig ... end

The symbolic interpreter using transducers