Module Colis.Language

The CoLiS language

module Nat = Colis__.Syntax__Nat

module Syntax = Colis__.Syntax__Syntax

module SyntaxHelpers = Colis__.SyntaxHelpers

module Parser = Colis__.ColisParser

module Lexer = Colis__.ColisLexer

module FromShell = Colis__.FromShell

Printing

val print_colis : colis -> unit

Prints a Colis program to stdout.

val colis_to_string : colis -> string

Prints a Colis program to a string.

val colis_to_file : string -> colis -> unit

Prints a Colis program to a file.

val pp_print_colis : Stdlib.Format.formatter -> colis -> unit

Generic pretty-printing function for Colis.

Parsing

val parse_colis_channel : ?⁠filename:string -> Stdlib.in_channel -> colis

Reads Colis syntax from a channel and returns the corresponding AST.

raises {!Errors.ParseError}

val parse_colis_file : string -> colis

Reads Colis syntax from a file and returns the corresponding AST.

raises {!Errors.ParseError}

val parse_colis_string : string -> colis

Reads Colis syntax from a string and returns the corresponding AST.

raises {!Errors.ParseError}

val convert_shell_file : cmd_line_arguments:string list -> Morsmall.AST.program -> colis

Converts the given Shell script to Colis and return the corresponding AST.

raises {!Errors.ConversionError}

val parse_shell_file : cmd_line_arguments:string list -> string -> colis

Reads Shell from a file, converts it to Colis and returns the corresponding AST.

raises {!Errors.ParseError}

raises {!Errors.ConversionError}

val embellish_colis : colis -> colis