Module Colis.SymbolicConstraints

The symbolic interpreter using constraints on the mixed backend of SymbolicUtility

type state = Colis__.SymbolicUtility.Constraints.state

type sym_state = Colis__.SymbolicUtility.Constraints.sym_state

type config = Colis__.SymbolicUtility.Constraints.config = {

prune_init_state : bool;
Prune the initial symbolic state during symbolic execution for a faster execution

}

val is_registered : name:string -> bool

Test if an utility is registerered in the mixed backend (the actual backend for this module)

val interp_program : loop_limit:int -> stack_size:int -> argument0:string -> sym_state list -> Language.Syntax.program -> state list * state list * state list

val run : config -> sym_config -> argument0:string -> ?⁠arguments:string list -> ?⁠vars:(string * string) list -> colis -> unit

val print_state : Stdlib.Format.formatter -> ?⁠id:string -> state -> unit

val print_states : initials:state list -> (state list * state list * state list) -> unit

For colis-batch

val add_fs_spec_to_clause : Colis_constraints_common.Var.t -> Constraints.Clause.sat_conj -> FilesystemSpec.t -> Constraints.Clause.sat_conj list

compile_fs_spec root conj fs_spec creates a disjunction that represents the conjunction conj with constraints representing the filesystem specified by fs_spec

val to_state : prune_init_state:bool -> root:Colis_constraints_common.Var.t -> Constraints.Clause.sat_conj -> state

val to_symbolic_state : vars:(string * string) list -> arguments:string list -> state -> sym_state