Module type Constraints.SYMBOLIC_UTILITY

A symbolic utility is comprised of a name and a function that interpretes a utility in a context.

val name : string

val interprete : Colis__.Semantics__UtilityContext.utility_context -> utility