Module SymbolicUtility.MakeInterpreter

Parameters

Filesystem : sig ... end

Signature

type filesystem = Filesystem.filesystem

Interpreter state

type state = {

filesystem : filesystem;

stdin : Colis__.Semantics__Buffers.Stdin.t;

stdout : Colis__.Semantics__Buffers.Stdout.t;

log : Colis__.Semantics__Buffers.Stdout.t;

}

val mk_state : filesystem -> state

Shortcut to make a state with empty stdin, stdout, and log

val print_stdout : newline:bool -> string -> state -> state

Print to stdout (and log)

val print_error : string -> state -> state

Print message as utility trace to log if it is not empty (marked as ERR)

val print_utility_trace : string -> state -> state

Print message as utility trace to log if it is not empty (marked as UTL)

val print_incomplete_trace : string -> state -> state

Print message as utility trace to log if it is not empty (marked as INCOMPLETE)

Utilities

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

A utility transforms a symbolic state into a list of symbolic states with boolean results

Basic utilities and behaviours

The symbolic interpretation of a utility call may result in different behaviours:

Any behaviour
  ├─ Known behaviour
  │  ├─ The call was interpreted
  │  │  ├─ [success]: The call was valid and succeeded (i.e. exit 0, i.e. Ok true)
  │  │  └─ [error]: The call failed (i.e. exit >0, i.e. Ok false)
  │  └─ [incomplete]: The utility or option has not been implemented (i.e. Incomplete)
  └─ [unknown]: We don't even know if this call is an error or incomplete behaviour

return true can be used for a simple success. error adds logging to return false.

A call is valid if it corresponds to the syntax of the utility.

The actual behaviour of unknown in the interpreter is determined by option Option.unknown_behaviour and may be an OCaml exception, or correspond to error behaviour, or incomplete behaviour.

val error : utility:string -> string -> utility

Error utility

val incomplete : utility:string -> string -> utility

Unsupported stuff in a known utility.

val unknown : utility:string -> string -> utility

Unknown-unknown behaviour *

Registration and dispatch

module type SYMBOLIC_UTILITY = sig ... end

A symbolic utility is comprised of a name and a function that interpretes a utility in a context.

val register : (module SYMBOLIC_UTILITY) -> unit

Register a symbolic utility

val is_registered : name:string -> bool

Check if a utility with name is registered

val dispatch : name:string -> Colis__.Semantics__UtilityContext.utility_context -> utility

Dispatch a call to utility name in a utility context

val call : string -> Colis__.Semantics__UtilityContext.utility_context -> string list -> utility

A saner way to call dispatch (e.g. in the implementation of utilities)

Interpretation

type sym_state = {

context : Colis__.Semantics__Context.context;

state : state;

}

A symbolic state combines a state with a concrete interpreter context

val interp_program : Colis__.Semantics__Input.input -> sym_state list -> Colis__.Syntax__Syntax.program -> state list * state list * state list

interp_program inp stas p symbolic interpretation configured by inp of a program p in a list of symbolic input states stas. Returns the symbolic states representing successful, errorneous, and incomplete interpretation.