Module SymbolicUtility.Mixed

Mixed constraints/transducers

The mixed backend combines the constraints-based backend with the transducers-based backend, while sharing the infrastructure for concrete evaluation.

type filesystem =

| Constraints of Constraints.filesystem

| Transducers of Transducers.filesystem

type case_spec

val case_spec : ?⁠transducers:Transducers.case_spec -> ?⁠constraints:Constraints.case_spec -> unit -> case_spec

A mixed case specification combines case specifications of the different backends. If the case specification is left unspecified for a backends, it induces incomplete behaviour for that backend independently of the specification case.

include INTERPRETER with type filesystem := filesystem

type filesystem

Interpreter state

type state = {

filesystem : filesystem;

stdin : Colis__.Semantics__Buffers.Stdin.t;

stdout : Colis__.Semantics__Buffers.Stdout.t;

log : Colis__.Semantics__Buffers.Stdout.t;

}

val mk_state : filesystem -> state

Shortcut to make a state with empty stdin, stdout, and log

val print_stdout : newline:bool -> string -> state -> state

Print to stdout (and log)

val print_error : string -> state -> state

Print message as utility trace to log if it is not empty (marked as ERR)

val print_utility_trace : string -> state -> state

Print message as utility trace to log if it is not empty (marked as UTL)

val print_incomplete_trace : string -> state -> state

Print message as utility trace to log if it is not empty (marked as INCOMPLETE)

Utilities

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

A utility transforms a symbolic state into a list of symbolic states with boolean results

Basic utilities and behaviours

The symbolic interpretation of a utility call may result in different behaviours:

Any behaviour
  ├─ Known behaviour
  │  ├─ The call was interpreted
  │  │  ├─ [success]: The call was valid and succeeded (i.e. exit 0, i.e. Ok true)
  │  │  └─ [error]: The call failed (i.e. exit >0, i.e. Ok false)
  │  └─ [incomplete]: The utility or option has not been implemented (i.e. Incomplete)
  └─ [unknown]: We don't even know if this call is an error or incomplete behaviour

return true can be used for a simple success. error adds logging to return false.

A call is valid if it corresponds to the syntax of the utility.

The actual behaviour of unknown in the interpreter is determined by option Option.unknown_behaviour and may be an OCaml exception, or correspond to error behaviour, or incomplete behaviour.

val error : utility:string -> string -> utility

Error utility

val incomplete : utility:string -> string -> utility

Unsupported stuff in a known utility.

val unknown : utility:string -> string -> utility

Unknown-unknown behaviour *

Registration and dispatch

module type SYMBOLIC_UTILITY = sig ... end

A symbolic utility is comprised of a name and a function that interpretes a utility in a context.

val register : (module SYMBOLIC_UTILITY) -> unit

Register a symbolic utility

val is_registered : name:string -> bool

Check if a utility with name is registered

val dispatch : name:string -> Colis__.Semantics__UtilityContext.utility_context -> utility

Dispatch a call to utility name in a utility context

val call : string -> Colis__.Semantics__UtilityContext.utility_context -> string list -> utility

A saner way to call dispatch (e.g. in the implementation of utilities)

Interpretation

type sym_state = {

context : Colis__.Semantics__Context.context;

state : state;

}

A symbolic state combines a state with a concrete interpreter context

val interp_program : Colis__.Semantics__Input.input -> sym_state list -> Colis__.Syntax__Syntax.program -> state list * state list * state list

interp_program inp stas p symbolic interpretation configured by inp of a program p in a list of symbolic input states stas. Returns the symbolic states representing successful, errorneous, and incomplete interpretation.

include COMBINATORS with type state := state

type state

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

val choice : utility -> utility -> utility

choice u1 u2 yields the utility that non-deterministacillay behaves like u1 or u2.

val return : bool -> utility

return b yields the utility that does not change its state, and succeeds if and only if b is true

val if_then_else : utility -> utility -> utility -> utility

if_then_else u1 u2 u3 yields the utility that behaves like if u1 then u2 else u3 fi

val if_then : utility -> utility -> utility

if_then u1 u2 yields the utility that behaves like if u1 then u2 fi

val uneg : utility -> utility

uneg u yields the utility that behaves like u but inverts its result

val uand : utility -> utility -> utility

uand u1 u2 yields the utility that behaves like u1 && u2 in a NON-LAZY manner

val uor : utility -> utility -> utility

uor u1 u2 yields the utility that behaves like u1 || u2 in a NON-LAZY manner

val multiple_times : ('a -> utility) -> 'a list -> utility

multiple_times what args executes what on every argument in args. It does not stop if one of these executions fails but the global utility fails in the case.

val compose_non_strict : utility -> utility -> utility

compose_non_strict u1 u2 yields the utility that behaves like u1; u2 in non-strict mode, that is the error code of u1 is ignored

val compose_strict : utility -> utility -> utility

compose_strict u1 u2 yields the utility that behaves like u1; u2 in strict mode, that is if u1 fails then the composition fails and u2 is not executed

Arguments Parsing

val cmdliner_eval_utility : utility:string -> ?⁠empty_pos_args:bool -> (Colis__.Semantics__UtilityContext.utility_context -> string list -> utility) Cmdliner.Term.t -> Colis__.Semantics__UtilityContext.utility_context -> utility

A wrapper around Cmdliner.Term.eval for utilities. utility is the name of the utility. empty_pos_args describe whether empty positional arguments lists are accepted or not (refused by default).

include SPECIFICATIONS with type state := state and type case_spec := case_spec

type state

type case_spec

val noop : case_spec

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

type case

A case in the specification is either a success, an error, or incomplete

val success_case : descr:string -> ?⁠stdout:Colis__.Semantics__Buffers.Stdout.t -> case_spec -> case

A success case (aka. return 0)

val error_case : descr:string -> ?⁠stdout:Colis__.Semantics__Buffers.Stdout.t -> ?⁠error_message:string -> case_spec -> case

An error case (aka return 1)

val incomplete_case : descr:string -> case_spec -> case

An incomplete case (unknown behaviour, cannot be symbolically executed)

val specification_cases : case list -> utility

Use a list of cases to specify a utility.

val interp_program_constraints : Colis__.Semantics__Input.input -> Constraints.sym_state list -> Colis__.Syntax__Syntax.program -> Constraints.state list * Constraints.state list * Constraints.state list

Symbolically interprete a program using the constraints backend

val interp_program_transducers : Colis__.Semantics__Input.input -> Transducers.sym_state list -> Colis__.Syntax__Syntax.program -> Transducers.state list * Transducers.state list * Transducers.state list

Symbolically interprete a program using the transducers backend

type utility_context = Colis__.Semantics__UtilityContext.utility_context = {

cwd : Colis_constraints_common.Path.normal;

env : string Colis__.Env.SMap.t;

args : string list;

}

Alias for easier access in the utility implementations.