Module Colis__FilesystemSpec

type t

Simple specification of filessytems (with a translation to feature constraints)

val empty : t

Empty filesystem (i.e., only the root directory)

val add_dir : string list -> t -> t

Add a directory to a FS spec at the given relative path to root, defined by a list of feature names.

raises Invalid_argument

if any super-path is already specified as a file.

val add_file : string list -> t -> t

Add a file to a FS spec at the given relative path to root.

raises Invalid_argument

if any super-path of the parent is specified as a file or the path is already specified as a directory.

val add_channel : Stdlib.in_channel -> t -> t

Adds files and directories from a channel. One line per entry, all lines start with '/' and directories end with '/'

raises Invalid_argument

like functions add_dir and add_file

val compile_constraints : Colis_constraints_common.Var.t -> t -> Colis_constraints.Clause.t

Compile the specification of a filesystem into a feature constraint at the given root variable.

val pp : Stdlib.Format.formatter -> t -> unit

Print the FS spec as a tree