Module Colis__SymbolicUtility.MakeCombinators

Parameters

Interpreter : INTERPRETER

Signature

type state = Interpreter.state

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

val choice : utility -> utility -> utility

choice u1 u2 yields the utility that non-deterministacillay behaves like u1 or u2.

val return : bool -> utility

return b yields the utility that does not change its state, and succeeds if and only if b is true

val if_then_else : utility -> utility -> utility -> utility

if_then_else u1 u2 u3 yields the utility that behaves like if u1 then u2 else u3 fi

val if_then : utility -> utility -> utility

if_then u1 u2 yields the utility that behaves like if u1 then u2 fi

val uneg : utility -> utility

uneg u yields the utility that behaves like u but inverts its result

val uand : utility -> utility -> utility

uand u1 u2 yields the utility that behaves like u1 && u2 in a NON-LAZY manner

val uor : utility -> utility -> utility

uor u1 u2 yields the utility that behaves like u1 || u2 in a NON-LAZY manner

val multiple_times : ('a -> utility) -> 'a list -> utility

multiple_times what args executes what on every argument in args. It does not stop if one of these executions fails but the global utility fails in the case.

val compose_non_strict : utility -> utility -> utility

compose_non_strict u1 u2 yields the utility that behaves like u1; u2 in non-strict mode, that is the error code of u1 is ignored

val compose_strict : utility -> utility -> utility

compose_strict u1 u2 yields the utility that behaves like u1; u2 in strict mode, that is if u1 fails then the composition fails and u2 is not executed

Arguments Parsing

val cmdliner_eval_utility : utility:string -> ?⁠empty_pos_args:bool -> (Colis__.Semantics__UtilityContext.utility_context -> string list -> utility) Cmdliner.Term.t -> Colis__.Semantics__UtilityContext.utility_context -> utility

A wrapper around Cmdliner.Term.eval for utilities. utility is the name of the utility. empty_pos_args describe whether empty positional arguments lists are accepted or not (refused by default).