Module type Colis__SymbolicUtility.SPECIFICATIONS

Specifications of a symbolic utility by specifications

type state

type case_spec

val noop : case_spec

type utility = state -> (state * bool Colis__.Semantics__Result.result) list

type case

A case in the specification is either a success, an error, or incomplete

val success_case : descr:string -> ?⁠stdout:Colis__.Semantics__Buffers.Stdout.t -> case_spec -> case

A success case (aka. return 0)

val error_case : descr:string -> ?⁠stdout:Colis__.Semantics__Buffers.Stdout.t -> ?⁠error_message:string -> case_spec -> case

An error case (aka return 1)

val incomplete_case : descr:string -> case_spec -> case

An incomplete case (unknown behaviour, cannot be symbolically executed)

val specification_cases : case list -> utility

Use a list of cases to specify a utility.