Module Colis_constraints_efficient_clause.Core

Core

type var

Type of the variables.

module VarSet : Stdlib.Set.S with type VarSet.elt = var

Information

module Info : sig ... end

Main Structure

type t

Type of an irreducible existential clause.

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val empty : t

Empty existential clause. That is, "true".

Variables and information

val hash : var -> int

val equal : var -> var -> t -> bool

val syntactic_compare : var -> var -> int

Comparison not in the structure but syntactically. Implies equality in the structure.

val identify : var -> var -> (Info.t -> Info.t -> Info.t) -> t -> t

Identify two variables. After that, they are "equal" in the structure. The function talking about infos is here to tell to the structure how to merge the info of the two variables.

val fresh_var : t -> var * t

val internalise : Colis_constraints_common.Var.t -> t -> var * t

val externalise : var -> t -> Colis_constraints_common.Var.t list

Returns the (possibly empty) list of external variables mapping to that particular (not up to equality) internal variable.

val fold_globals : (var -> 'a -> 'a) -> t -> 'a -> 'a

val quantify_over : Colis_constraints_common.Var.t -> t -> t

val is_shadow : var -> t -> bool

val get_info : var -> t -> Info.t

get_info x c returns the information associated with variable x in clause c, after having updated the shadow field.

val get_info_no_shadow : var -> t -> Info.t

Same as get_info but does not update the shadow field.

val set_info : var -> t -> Info.t -> t

val update_info : var -> t -> (Info.t -> Info.t) -> t

Equivalent to get_info x c |> f |> set_info x c.

val filter_infos : (var -> bool) -> t -> t

val fold_infos : (var -> Info.t -> 'a -> 'a) -> t -> 'a -> 'a

fold_infos f c e computes f in order for all variables and associated informations in the all clause c, passing the result to the next computation. This does not update the shadow field.

val iter_infos : (var -> Info.t -> unit) -> t -> unit

val iter_equalities : (var -> var -> unit) -> t -> unit

sims

val remove_sim : var -> t -> Info.t -> Info.t

remove_sim y c info removes all references to variables equal to y (wrt. structure c) in info.

val update_sim : var -> (Colis_constraints_common.Feat.Set.t option -> Colis_constraints_common.Feat.Set.t) -> t -> Info.t -> Info.t

update_sim y f c info updates the sims in info by calling f None if there is no sim for y or Some gs if there is a sim ~gs y.

val update_info_for_all_similarities : (Colis_constraints_common.Feat.Set.t -> var -> Info.t -> Info.t) -> var -> Info.t -> t -> t

update_info_for_all_similarities upd x info takes a clause and applies the upd function to all the info records of variables that are similar to the given info (including it).

nsims

val remove_nsims : var -> t -> t

Remove all nsims in the given info.

Garbage Collection

val simplify : t -> t

Removes all variables that are not accessible from the globals.

val with_shadow_variables : (unit -> 'a) -> 'a

with_initial f runs f in an environment where all new variables are declared to be shadow.

Should Disappear

FIXME

val not_implemented : string -> 'a

val pp_debug : Stdlib.Format.formatter -> t -> unit